requirements.txt
google-adk==1.23.0
python-telegram-bot==22.6
telegramify-markdown==0.5.4
logfire==4.22.0
jinja2==3.1.6